Split rhea into rhea,rdf
